Fix it cannot be compiled for Windows CE.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@881 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/ui/textwindow.cpp
@@ -1012,7 +1012,7 @@ std::pair<unsigned char, unsigned char> qs::TextWindowImpl::getLineQuoteDepth(co
 	const WCHAR* p = line.getText();
 	size_t nLength = line.getLength();
 	if (nLength != 0 && wcschr(wstrQuote_[0].get(), *p)) {
-		size_t nMax = QSMIN(nLength, MAX_QUOTE);
+		size_t nMax = QSMIN(nLength, size_t(MAX_QUOTE));
 		const WCHAR* pLastQuote = p;
 		for (size_t n = 0; n < nMax; ++n, ++p) {
 			if (wcschr(wstrQuote_[0].get(), *p)) {
